+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - `ToString` trait implementation for validators.
 - Define `JSONSchema::options` to customise `JSONSchema` compilation [#131](https://github.com/Stranger6667/jsonschema-rs/issues/131)
+- Allow user-defined `contentEncoding` and `contentMediaType` keywords
 
 ### Fixed